@@ -3,17 +3,17 @@ description: A new flutter plugin project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   async: ^2.6.1
-#  path: ^1.8.0
-#  pedantic: ^1.11.0
+  #  path: ^1.8.0
+  #  pedantic: ^1.11.0
   encrypt: ^5.0.0
-  xml2json: ^5.2.0
-#  puppeteer: ^2.0.0
+  xml2json: ^5.3.1
+#  puppeteer: ^2.2.0
   uuid: 3.0.4
 
 
