@@ -9,7 +9,7 @@ class BaiduMusicPage extends StatefulWidget {
 }
 
 class _BaiduMusicPageState extends State<BaiduMusicPage> with AutomaticKeepAliveClientMixin {
-  BaiduRepository baidu;
+  BaiduRepository? baidu;
   String result = "";
 
   @override
@@ -45,109 +45,109 @@ class _BaiduMusicPageState extends State<BaiduMusicPage> with AutomaticKeepAlive
                 ListTile(
                   title: Text('开屏广告(可能为空)'),
                   onTap: () {
-                    baidu.openScreen().then(onData).catchError(onError);
+                    baidu?.openScreen().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('首页'),
                   onTap: () {
-                    baidu.index().then(onData).catchError(onError);
+                    baidu?.index().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('专辑列表(最新专辑)'),
                   onTap: () {
-                    baidu.albumList().then(onData).catchError(onError);
+                    baidu?.albumList().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('专辑详情'),
                   onTap: () {
-                    baidu.albumInfo().then(onData).catchError(onError);
+                    baidu?.albumInfo().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('歌曲列表(新歌推荐)'),
                   onTap: () {
-                    baidu.songList().then(onData).catchError(onError);
+                    baidu?.songList().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('歌曲详情'),
                   onTap: () {
-                    baidu.songInfo().then(onData).catchError(onError);
+                    baidu?.songInfo().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('歌曲下载'),
                   onTap: () {
-                    baidu.songDownload().then(onData).catchError(onError);
+                    baidu?.songDownload().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('歌手列表'),
                   onTap: () {
-                    baidu.artistList().then(onData).catchError(onError);
+                    baidu?.artistList().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('歌手信息'),
                   onTap: () {
-                    baidu.artistInfo().then(onData).catchError(onError);
+                    baidu?.artistInfo().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('歌手歌曲'),
                   onTap: () {
-                    baidu.artistSong().then(onData).catchError(onError);
+                    baidu?.artistSong().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('歌手专辑'),
                   onTap: () {
-                    baidu.artistAlbum().then(onData).catchError(onError);
+                    baidu?.artistAlbum().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('搜索'),
                   onTap: () {
-                    baidu.search().then(onData).catchError(onError);
+                    baidu?.search().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('搜索热词'),
                   onTap: () {
-                    baidu.searchSug().then(onData).catchError(onError);
+                    baidu?.searchSug().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('榜单类型'),
                   onTap: () {
-                    baidu.bdType().then(onData).catchError(onError);
+                    baidu?.bdType().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('榜单歌曲列表'),
                   onTap: () {
-                    baidu.bdList().then(onData).catchError(onError);
+                    baidu?.bdList().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('歌单分类'),
                   onTap: () {
-                    baidu.tracklistType().then(onData).catchError(onError);
+                    baidu?.tracklistType().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('歌单列表'),
                   onTap: () {
-                    baidu.tracklistList().then(onData).catchError(onError);
+                    baidu?.tracklistList().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('歌单详情'),
                   onTap: () {
-                    baidu.tracklistInfo().then(onData).catchError(onError);
+                    baidu?.tracklistInfo().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
@@ -156,25 +156,25 @@ class _BaiduMusicPageState extends State<BaiduMusicPage> with AutomaticKeepAlive
                 ListTile(
                   title: Text('视频列表'),
                   onTap: () {
-                    baidu.videoList().then(onData).catchError(onError);
+                    baidu?.videoList().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('视频详情'),
                   onTap: () {
-                    baidu.videoInfo().then(onData).catchError(onError);
+                    baidu?.videoInfo().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('视频推荐'),
                   onTap: () {
-                    baidu.videoRecommend().then(onData).catchError(onError);
+                    baidu?.videoRecommend().then(onData).catchError(onError);
                   },
                 ),
                 ListTile(
                   title: Text('视频下载'),
                   onTap: () {
-                    baidu.videoDownload().then(onData).catchError(onError);
+                    baidu?.videoDownload().then(onData).catchError(onError);
                   },
                 ),
               ],
