@@ -26,29 +26,29 @@ class NetRepository {
     return jar.loadForRequest(uri);
   }
 
-  _saveCookies(List<Cookie> cookies) async {
+  _saveCookies(List<Cookie>? cookies) async {
     var jar = await _cookieJar.future;
-    if (jar == null) {
+    if (jar == null || cookies == null) {
       return;
     }
     final uri = Uri.parse('http://music.163.com');
     jar.saveFromResponse(uri, cookies);
   }
 
-  Future<dynamic> _doRequest(String path, {Map<String, dynamic> params}) async {
+  Future<dynamic> _doRequest(String path, {Map<String, dynamic>? params}) async {
     List<Cookie> cookies = await _loadCookies();
-    Answer answer;
+    Answer? answer;
     try {
       answer = await cloudMusicApi(path, parameter: params, cookie: cookies);
     } catch (e, stacktrace) {
       Future.error('request error:$e', stacktrace);
     }
 
-    if (answer.status == HttpStatus.ok) {
-      _saveCookies(answer.cookie);
+    if (answer?.status == HttpStatus.ok) {
+      _saveCookies(answer?.cookie);
     }
 
-    final map = answer.body;
+    final map = answer?.body;
     if (map == null) {
       return Future.error('请求失败了');
     } else if (map['code'] == _NEED_LOGIN) {
