@@ -13,8 +13,8 @@ DebugPrinter debugPrint = (msg) {
 /// 百度音乐 API
 Future<dynamic> baiduMusicApi(
   String path, {
-  Map parameter,
-  String auth,
+  Map? parameter,
+  String? auth,
 }) async {
   // assert(path != null, "path can not be null");
   // assert(handles.containsKey(path), "此 api url 未被定义, 请检查: $path ");
@@ -23,10 +23,10 @@ Future<dynamic> baiduMusicApi(
         .copy(body: {'code': 500, 'msg': "此 api url 未被定义, 请检查: $path ",'path':handles.keys.toList()});
   }
 
-  final Handler handle = handles[path];
+  final Handler? handle = handles[path];
 
   try {
-    final answer = await handle(parameter, auth);
+    final answer = await handle!(parameter??{}, auth);
     return answer;
   } on HttpException catch (e, stack) {
     debugPrint(e.toString());
