@@ -13,7 +13,7 @@ DebugPrinter debugPrint = (msg) {
 /// migu音乐 API
 Future<dynamic> miguApi(
   String path, {
-  Map parameter,
+  Map? parameter,
   List<Cookie> cookie = const [],
 }) async {
   // assert(path != null, "path can not be null");
@@ -21,10 +21,10 @@ Future<dynamic> miguApi(
   if (!handles.containsKey(path)) {
     return Answer().copy(body: {'code': 500, 'msg': "此 api url 未被定义, 请检查: $path ",'path':handles.keys.toList()});
   }
-  final Handler handle = handles[path];
+  final Handler? handle = handles[path];
 
   try {
-    final answer = await handle(parameter, cookie);
+    final answer = await handle!(parameter??{}, cookie);
     return answer;
   } on HttpException catch (e, stack) {
     debugPrint(e.toString());
