@@ -33,7 +33,7 @@ Handler userBindingCellphone = (query, cookie) {
     'phone': query['phone'],
     'countrycode': query['countrycode'] ?? '86',
     'captcha': query['captcha'],
-    'password': Encrypted(md5.convert(utf8.encode(query['password'])).bytes).base16,
+    'password': Encrypted(Uint8List.fromList(md5.convert(utf8.encode(query['password'])).bytes)).base16,
   };
 
   return request(
