@@ -14,7 +14,7 @@ DebugPrinter debugPrint = (msg) {
 /// API文档地址参考: https://binaryify.github.io/NeteaseCloudMusicApi/#/
 Future<Answer> cloudMusicApi(
   String path, {
-  Map parameter,
+  Map? parameter,
   List<Cookie> cookie = const [],
 }) async {
   // assert(path != null, "path can not be null");
@@ -23,10 +23,10 @@ Future<Answer> cloudMusicApi(
     return Answer()
         .copy(body: {'code': 500, 'msg': "此 api url 未被定义, 请检查: $path ",'path':handles.keys.toList()});
   }
-  final Handler handle = handles[path];
+  final Handler? handle = handles[path];
 
   try {
-    final answer = await handle(parameter, cookie);
+    final answer = await handle!(parameter??{}, cookie);
     return answer;
   } on HttpException catch (e, stack) {
     debugPrint(e.toString());
