@@ -7,7 +7,7 @@ class Answer {
 
   const Answer({this.status = 500, this.body = const {'code': 500, 'msg': 'server error'}, this.cookie = const []});
 
-  Answer copy({int status, Map body, List cookie}) {
+  Answer copy({int? status, Map? body, List<Cookie>? cookie}) {
     return Answer(status: status ?? this.status, body: body ?? this.body, cookie: cookie ?? this.cookie);
   }
 }
